requirements/base.txt
numpy
scipy
pandas

torch
transformers
sentence-transformers

faiss-cpu
rank-bm25

tqdm
rich
pyyaml